django_project/pyproject.toml
[tool.poetry]
name = "django_project"
version = "0.1.0"
description = "Django backend for NFT Gateway"
authors = ["Roger Olowski <rogerolowski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.16"
django-tailwind = "^3.7"
psycopg2-binary = "^2.9"
redis = "^5.0"
django-redis = "^5.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
